+++ apps/h5/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.63](https://github.com/samchen08/vtj.pro/compare/vtj-project-h5@0.12.62...vtj-project-h5@0.12.63) (2025-07-19)
+
+**Note:** Version bump only for package vtj-project-h5
+
+
+
+
+
+## [0.12.62](https://github.com/samchen08/vtj.pro/compare/vtj-project-h5@0.12.61...vtj-project-h5@0.12.62) (2025-07-19)
+
+**Note:** Version bump only for package vtj-project-h5
+
+
+
+
+
+## [0.12.61](https://github.com/samchen08/vtj.pro/compare/vtj-project-h5@0.12.60...vtj-project-h5@0.12.61) (2025-07-19)
+
+**Note:** Version bump only for package vtj-project-h5
+
+
+
+
+
 ## [0.12.60](https://github.com/samchen08/vtj.pro/compare/vtj-project-h5@0.12.59...vtj-project-h5@0.12.60) (2025-07-18)
 
 **Note:** Version bump only for package vtj-project-h5