+++ packages/coder/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.63](https://gitee.com/newgateway/vtj/compare/@vtj/coder@0.12.62...@vtj/coder@0.12.63) (2025-07-19)
+
+**Note:** Version bump only for package @vtj/coder
+
+
+
+
+
+## [0.12.62](https://gitee.com/newgateway/vtj/compare/@vtj/coder@0.12.61...@vtj/coder@0.12.62) (2025-07-19)
+
+
+### Bug Fixes
+
+* 🐛 coder 双引号问题 ([bec5a28](https://gitee.com/newgateway/vtj/commits/bec5a28bf8404cebc73ce6747c090a175d40eea3))
+
+
+
+
+
+## [0.12.61](https://gitee.com/newgateway/vtj/compare/@vtj/coder@0.12.60...@vtj/coder@0.12.61) (2025-07-19)
+
+**Note:** Version bump only for package @vtj/coder
+
+
+
+
+
 ## [0.12.60](https://gitee.com/newgateway/vtj/compare/@vtj/coder@0.12.59...@vtj/coder@0.12.60) (2025-07-18)